BoardConfig: Handling of MR0 blobs and testing for fix the flickering :)
WE NEED TO SOLVE DAT PROBLEM
DisplayDevice: Backwards compatibility with old EGL

From 4.1 to 4.2, the display subsystem was reworked to
use SurfaceTextureClient/BufferQueue instead of
FramebufferNativeWindow for the framebuffer itself.
Unfortunately, some legacy EGL libraries make assumptions
that any framebuffer device will be FramebufferNativeWindow.
These EGL libraries will fail when used in 4.2 as if the
framebuffer is not FramebufferNativeWindow, they will
try to dequeue more than one buffer at a time, which
will cause a hang of the graphics subsystem
http://review.cyanogenmod.org/#/c/28005/

Also, we disable triple buffering for now, since it's the handler of the bad-done VSYNC, so, we cap the FPS rate by refresh_rate cflag to 60, and let the screen do the actual capping of Hertz screen refresh rate. As explained here, the triple framebuffering is NOT checking the actual refresh rate of the screen, so it's running as fast as it can :)

Due to the software algorithm not having to poll the graphics hardware for monitor refresh events, the algorithm is free to run as fast as possible.